pyproject.toml
[project]
name = "gemeni-pdf-to-markdown"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.8.3",
    "pdf2image>=1.17.0",
    "pillow>=11.0.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.0.1",
    "pytest>=8.3.4",
    "ruff>=0.8.3",
]
